+++ examples/04_abort/src/Person.tsx
@@ -54,16 +54,16 @@ const reducer: Reducer<State, Action> = (state, action) => {
 type AsyncAction = { type: 'FETCH_PERSON'; id: number }
 
 const asyncActionHandlers: AsyncActionHandlers<Reducer<State, Action>, AsyncAction> = {
-  FETCH_PERSON: ({ dispatch }) => async (action) => {
+  FETCH_PERSON: ({ dispatch, signal }) => async (action) => {
     dispatch({ type: 'START_FETCH' });
     try {
-      const response = await fetch(`https://reqres.in/api/users/${action.id}?delay=1`);
+      const response = await fetch(`https://reqres.in/api/users/${action.id}?delay=1`, { signal });
       const data = await response.json();
       const firstName = data.data.first_name;
       if (typeof firstName !== 'string') throw new Error();
-      dispatch({ type: 'FINISH_FETCH', firstName });
+      if (!signal.aborted) dispatch({ type: 'FINISH_FETCH', firstName });
     } catch (e) {
-      dispatch({ type: 'ERROR_FETCH' });
+      if (!signal.aborted) dispatch({ type: 'ERROR_FETCH' });
     }
   },
 };