+++ src/index.ts
@@ -19,16 +19,28 @@ const isClient = (
 
 const useIsomorphicLayoutEffect = isClient ? useLayoutEffect : useEffect;
 
+const useAbortSignal = () => {
+  const abortController = useRef(new AbortController());
+  useEffect(() => {
+    const abort = () => {
+      abortController.current.abort();
+    };
+    return abort;
+  }, []);
+  return abortController.current.signal;
+};
+
 export type AsyncActionHandlers<
   R extends Reducer<any, any>,
   AsyncAction extends { type: string }
 > = {
   [T in AsyncAction['type']]: AsyncAction extends infer A ? A extends {
     type: T;
-  } ? (
-    dispatch: Dispatch<ReducerAction<R>>,
-    getState: () => ReducerState<R>,
-  ) => (a: A) => Promise<void> : never : never;
+  } ? (s: {
+    dispatch: Dispatch<ReducerAction<R>>;
+    getState: () => ReducerState<R>;
+    signal: AbortSignal;
+  }) => (a: A) => Promise<void> : never : never;
 };
 
 export function useReducerAsync<
@@ -49,12 +61,12 @@ export function useReducerAsync<
  * import { useReducerAsync } from 'use-reducer-async';
  *
  * const asyncActionHandlers = {
- *   SLEEP: (dispatch, getState) => async (action) => {
+ *   SLEEP: ({ dispatch, getState, signal }) => async (action) => {
  *     dispatch({ type: 'START_SLEEP' });
  *     await new Promise(r => setTimeout(r, action.ms));
  *     dispatch({ type: 'END_SLEEP' });
  *   },
- *   FETCH: (dispatch, getState) => async (action) => {
+ *   FETCH: ({ dispatch, getState, signal }) => async (action) => {
  *     dispatch({ type: 'START_FETCH' });
  *     try {
  *       const response = await fetch(action.url);
@@ -88,10 +100,11 @@ export function useReducerAsync<
   initializer: unknown,
   asyncActionHandlers?: AsyncActionHandlers<R, AsyncAction>,
 ): [ReducerState<R>, Dispatch<ExportAction>] {
+  const signal = useAbortSignal();
   const aaHandlers = (
     asyncActionHandlers || initializer
   ) as AsyncActionHandlers<R, AsyncAction>;
-  const [state, rawDispatch] = useReducer(
+  const [state, dispatch] = useReducer(
     reducer,
     initializerArg as any,
     asyncActionHandlers && initializer as any,
@@ -101,19 +114,20 @@ export function useReducerAsync<
     lastState.current = state;
   }, [state]);
   const getState = useCallback((() => lastState.current), []);
-  const dispatch = useCallback((action: AsyncAction | ReducerAction<R>) => {
+  const wrappedDispatch = useCallback((action: AsyncAction | ReducerAction<R>) => {
     const { type } = (action || {}) as { type?: AsyncAction['type'] };
     const aaHandler = (
       (type && aaHandlers[type]) || null
-    ) as (typeof action extends AsyncAction ? (
-      dispatch: Dispatch<ReducerAction<R>>,
-      getState: () => ReducerState<R>,
-    ) => (a: typeof action) => Promise<void> : null);
+    ) as (typeof action extends AsyncAction ? (s: {
+      dispatch: Dispatch<ReducerAction<R>>;
+      getState: () => ReducerState<R>;
+      signal: AbortSignal;
+    }) => (a: typeof action) => Promise<void> : null);
     if (aaHandler) {
-      aaHandler(rawDispatch, getState)(action as AsyncAction);
+      aaHandler({ dispatch, getState, signal })(action as AsyncAction);
     } else {
-      rawDispatch(action as ReducerAction<R>);
+      dispatch(action as ReducerAction<R>);
     }
-  }, [aaHandlers, getState]);
-  return [state, dispatch];
+  }, [aaHandlers, getState, signal]);
+  return [state, wrappedDispatch];
 }